+++ CHANGELOG.rst
@@ -23,6 +23,7 @@ Added
     - Sample sheet version export (#739)
     - Sample sheet version restoring (#701)
     - Save and restore sheet configuration with ``ISATab`` version
+    - Deletion of ``ISATab`` versions on sheet delete (#746)
 
 Changed
 -------